Cache built transaction
The problem was that whenever you call one of the operations requiring
built transaction (info, split, raw), build() method of the
TransactionBuilder is called. And transaction is rebuilt again.

It is not only a problem of performance. Since SecureRandom is used
while signing transaction it might (and most likely will) produce
signatures of different lengths between calls.

For instance, you call [info].
It shows that transaction is 100 byte and fee is 1 s/b.
Then you want to see raw transaction.
You call [raw]
At this time, transaction is rebuild and it is not the same transaction
for which info was shown.
This transaction might be a couple of bytes shorter (which is good) or
longer (which is not so good).

How the problem is solved and you get the same transaction between
raw, split and info calls. Resulting transaction is cached until you
execute some command altering builder.